Cache all GitHub and Artifact Registry API calls
In our typical monorepo-based usage, we're asking
"what sha is main in the monorepo" many many
times. And if multiple installations are at the
same version (eg dev and staging) or use the same
image, we're asking the same questions there too.

Use non-configurable lru-cache with 1024 entries
each for now.