update-git-refs: don't error if current or Docker ref unknown (#43)
We only try to resolve the current `ref` and the commit found in the
Docker tag as a heuristic to choose something other than the current
commit SHA at the tracked ref that might be better and equivalent on the
given subpath. So if either of them fail to resolve, we can ignore it
rather than failing the whole update operation.

Fixes #40.