update-git-refs: don't error if current/Docker has no tree path
This is similar to #43: we shouldn't treat an inability to understand
the current value we're trying to overwrite or the heuristic value we
get from the Docker tag as a hard error.

Before #46, this error usually wouldn't happen, because if the given
commit doesn't exist in the repo, the resolveRefToSHA would fail (and be
ignored by the #43 fix) and we'd never get to getTreeSHAForPath. But #46
means that we are often skipping the API call for resolveRefToSHA if it
already looks like a SHA. So now we have to ignore the error here too.