Remove some recently-added outputs
The relevant-commits output is part of an unfinished feature and we can
re-add it as we finish the feature; it also shouldn't be set repeatedly
on each file but only once at the top with merged contents.

The other outputs were a bit hardcoded for our particular repository and
we've found a way to get that information directly in our workflow
rather than through this action. (Additionally, these outputs should
have been set at the top level rather than in processFile as they are
the same for each file.)