Include Helm chart changes in commit PR description
In #60 we looked at Docker tag changes, and based on the git commits
mentioned in the tags and an analysis of all tags in the Artifact
Registry repository, we figured out what commits affected the Docker tag
and included information about it in the PR description.

Now we're doing the same thing for changes to the "ref" (Helm chart
changes).

The basic algorithm is simple: ask GitHub for the list of commits
affecting the Chart's path at the old and new ref; find the most recent
"old" commit in the "new" list, and keep anything newer.

I did a fair amount of refactoring to create data structures that are
explicit about the reasons that we might be changing a line in the file
without actually knowing the list of commits.

This also removes an older unused attempt at implementing something
similar (`compareCommits`).

Since this changes the format of the "git commits between tags" part of
the API cache, I renamed the relevant field so it gets loaded properly
(old caches will just be ignored)