Allow promotes to include named links
The text and URL of the link come from a provided template file.

For now, the only variable that can be passed to the template is
docker-image-sha256-63.

(We have designed this specifically so that database migration
promotions can link to GCP logs for a k8s job that ran `flyway info`.)