Enhance PR Description for cleanup mode (#94)
This change adds the team name, app name, and environment and groups PRs under them. I also added a closing time for the PR to see when it was closed at a glance.

Additionally I fixed a bug where the github URL was being rendered with an extra .git that broke web links.