Switch bundler from ncc to rollup
The typescript-action repo we forked did this when they switched from
CJS to ESM output. We want to treat input as ESM to support new ESM-only
modules. But `google-gax` (a dependency of
`@google-cloud/artifact-registry`) refers to `__dirname` so it really
can't be in ESM. So we use rollup, which can treat input as ESM but
still output to CJS.

Plus we have source maps now!