[FOUN-1002] Add promotion metadata to a comment in PR descriptions (#120)
For now, it just lists the source and target app for every non-no-op promotion, but there will probably be more metadata later. Using the ArkType library to define the schema so it's easy to validate in the app that will read it.

The first commit in this PR moves the tests from `__tests__` to `src/__tests__`, which (along with some other adjustments) means that CI (and `npx tsc`) typechecks the tests too. It also means we don't need a separate tsconfig file for eslint to find the tests.